Day 9

+++ src/puzzles/puzzle02.rs
@@ -3,13 +3,13 @@ use crate::intcode::IntcodeMachine;
 pub fn run(input: String) -> Vec<String> {
     let mut answers: Vec<String> = Vec::new();
 
-    let inputs: Vec<i32> = input.trim().split(',').map(|s| s.parse::<i32>().expect(&format!("Not a number found in input `{}`", s))).collect();
+    let inputs: Vec<i64> = input.trim().split(',').map(|s| s.parse::<i64>().expect(&format!("Not a number found in input `{}`", s))).collect();
 
     answers.push(format!("{}", calc(&mut inputs.clone(), 12, 2)));
 
-    let mut try_noun: i32 = 0;
-    let mut try_verb: i32 = 0;
-    let mut last_result: i32 = calc(&mut inputs.clone(), try_noun, try_verb);
+    let mut try_noun: i64 = 0;
+    let mut try_verb: i64 = 0;
+    let mut last_result: i64 = calc(&mut inputs.clone(), try_noun, try_verb);
     while last_result != 19690720 {
         if try_verb != 99 {
             try_verb += 1;
@@ -25,7 +25,7 @@ pub fn run(input: String) -> Vec<String> {
     answers
 }
 
-fn calc(state: &mut Vec<i32>, noun: i32, verb: i32) -> i32 {
+fn calc(state: &mut Vec<i64>, noun: i64, verb: i64) -> i64 {
     state[1] = noun;
     state[2] = verb;
