+++ src/puzzles/puzzle07.rs
@@ -5,15 +5,15 @@ use itertools::Itertools;
 pub fn run(input: String) -> Vec<String> {
     let mut answers: Vec<String> = Vec::new();
 
-    let program: Vec<i32> = input.trim().split(',').map(|s| s.parse::<i32>().expect(&format!("Not a number found in input `{}`", s))).collect();
+    let program: Vec<i64> = input.trim().split(',').map(|s| s.parse::<i64>().expect(&format!("Not a number found in input `{}`", s))).collect();
 
     answers.push(format!("{}", (0..=4).permutations(5).map(
         |test_input| {
             // First input is 0, which is taken from outputs, so put that here
-            let outputs: &mut Vec<i32> = &mut vec![0];
+            let outputs: &mut Vec<i64> = &mut vec![0];
 
             for i in test_input {
-                let inputs: &mut Vec<i32> = &mut outputs.clone();
+                let inputs: &mut Vec<i64> = &mut outputs.clone();
                 inputs.push(i);
                 outputs.clear();
                 IntcodeMachine::new(&mut program.clone(), inputs, outputs, 0).execute_until_halt();
@@ -25,11 +25,11 @@ pub fn run(input: String) -> Vec<String> {
 
     let mut threads = Vec::new();
     (5..=9).permutations(5).for_each(|test_input| {
-        let program: Vec<i32> = input.trim().split(',').map(|s| s.parse::<i32>().expect(&format!("Not a number found in input `{}`", s))).collect();
-        threads.push(thread::spawn(move || -> i32 {
-            let inputs: [&mut Vec<i32>; 5] = [&mut vec![0, test_input[0]], &mut vec![test_input[1]], &mut vec![test_input[2]], &mut vec![test_input[3]], &mut vec![test_input[4]]];
-            let outputs: [&mut Vec<i32>; 5] = [&mut vec![], &mut vec![], &mut vec![], &mut vec![], &mut vec![]];
-            let states: [&mut Vec<i32>; 5] = [&mut program.clone(), &mut program.clone(), &mut program.clone(), &mut program.clone(), &mut program.clone()];
+        let program: Vec<i64> = input.trim().split(',').map(|s| s.parse::<i64>().expect(&format!("Not a number found in input `{}`", s))).collect();
+        threads.push(thread::spawn(move || -> i64 {
+            let inputs: [&mut Vec<i64>; 5] = [&mut vec![0, test_input[0]], &mut vec![test_input[1]], &mut vec![test_input[2]], &mut vec![test_input[3]], &mut vec![test_input[4]]];
+            let outputs: [&mut Vec<i64>; 5] = [&mut vec![], &mut vec![], &mut vec![], &mut vec![], &mut vec![]];
+            let states: [&mut Vec<i64>; 5] = [&mut program.clone(), &mut program.clone(), &mut program.clone(), &mut program.clone(), &mut program.clone()];
             let mut positions: [usize; 5] = [0, 0, 0, 0, 0];
 
             // Run all the machines once to consume their initial state
